+++ Runtime/ARCoreAnchorSubsystem.cs
@@ -11,6 +11,10 @@ namespace UnityEngine.XR.ARCore
     [Preserve]
     public sealed class ARCoreAnchorSubsystem : XRAnchorSubsystem
     {
+        /// <summary>
+        /// Creates an instance of the ARCore-specific anchor provider. This provides anchor services for ARCore.
+        /// </summary>
+        /// <returns>A new instance of <see cref="Provider"/> specific to ARCore.</returns>
         protected override Provider CreateProvider() => new ARCoreProvider();
 
         class ARCoreProvider : Provider