+++ Runtime/ARCoreLoader.cs
@@ -4,90 +4,76 @@ using UnityEngine.XR.Management;
 
 namespace UnityEngine.XR.ARCore
 {
+    /// <summary>
+    /// Manages the lifecycle of ARCore subsystems.
+    /// </summary>
     public class ARCoreLoader : XRLoaderHelper
     {
-        private static List<XRSessionSubsystemDescriptor> s_SessionSubsystemDescriptors = new List<XRSessionSubsystemDescriptor>();
-        private static List<XRCameraSubsystemDescriptor> s_CameraSubsystemDescriptors = new List<XRCameraSubsystemDescriptor>();
-        private static List<XRDepthSubsystemDescriptor> s_DepthSubsystemDescriptors = new List<XRDepthSubsystemDescriptor>();
-        private static List<XRPlaneSubsystemDescriptor> s_PlaneSubsystemDescriptors = new List<XRPlaneSubsystemDescriptor>();
-        private static List<XRAnchorSubsystemDescriptor> s_AnchorSubsystemDescriptors = new List<XRAnchorSubsystemDescriptor>();
-        private static List<XRRaycastSubsystemDescriptor> s_RaycastSubsystemDescriptors = new List<XRRaycastSubsystemDescriptor>();
-        private static List<XRImageTrackingSubsystemDescriptor> s_ImageTrackingSubsystemDescriptors = new List<XRImageTrackingSubsystemDescriptor>();
-        private static List<XRInputSubsystemDescriptor> s_InputSubsystemDescriptors = new List<XRInputSubsystemDescriptor>();
-        private static List<XRFaceSubsystemDescriptor> s_FaceSubsystemDescriptors = new List<XRFaceSubsystemDescriptor>();
+        static List<XRSessionSubsystemDescriptor> s_SessionSubsystemDescriptors = new List<XRSessionSubsystemDescriptor>();
+        static List<XRCameraSubsystemDescriptor> s_CameraSubsystemDescriptors = new List<XRCameraSubsystemDescriptor>();
+        static List<XRDepthSubsystemDescriptor> s_DepthSubsystemDescriptors = new List<XRDepthSubsystemDescriptor>();
+        static List<XRPlaneSubsystemDescriptor> s_PlaneSubsystemDescriptors = new List<XRPlaneSubsystemDescriptor>();
+        static List<XRAnchorSubsystemDescriptor> s_AnchorSubsystemDescriptors = new List<XRAnchorSubsystemDescriptor>();
+        static List<XRRaycastSubsystemDescriptor> s_RaycastSubsystemDescriptors = new List<XRRaycastSubsystemDescriptor>();
+        static List<XRImageTrackingSubsystemDescriptor> s_ImageTrackingSubsystemDescriptors = new List<XRImageTrackingSubsystemDescriptor>();
+        static List<XRInputSubsystemDescriptor> s_InputSubsystemDescriptors = new List<XRInputSubsystemDescriptor>();
+        static List<XRFaceSubsystemDescriptor> s_FaceSubsystemDescriptors = new List<XRFaceSubsystemDescriptor>();
+        static List<XREnvironmentProbeSubsystemDescriptor> s_EnvironmentProbeSubsystemDescriptors = new List<XREnvironmentProbeSubsystemDescriptor>();
 
-        public XRSessionSubsystem sessionSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRSessionSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRSessionSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRSessionSubsystem sessionSubsystem => GetLoadedSubsystem<XRSessionSubsystem>();
 
-        public XRCameraSubsystem cameraSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRCameraSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRCameraSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRCameraSubsystem cameraSubsystem => GetLoadedSubsystem<XRCameraSubsystem>();
 
-        public XRDepthSubsystem depthSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRDepthSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRDepthSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRDepthSubsystem depthSubsystem => GetLoadedSubsystem<XRDepthSubsystem>();
 
-        public XRPlaneSubsystem planeSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRPlaneSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRPlaneSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRPlaneSubsystem planeSubsystem => GetLoadedSubsystem<XRPlaneSubsystem>();
 
-        public XRAnchorSubsystem anchorSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRAnchorSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRAnchorSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRAnchorSubsystem anchorSubsystem => GetLoadedSubsystem<XRAnchorSubsystem>();
 
-        public XRRaycastSubsystem raycastSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRRaycastSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRRaycastSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRRaycastSubsystem raycastSubsystem => GetLoadedSubsystem<XRRaycastSubsystem>();
 
-        public XRImageTrackingSubsystem imageTrackingSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRImageTrackingSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRImageTrackingSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRImageTrackingSubsystem imageTrackingSubsystem => GetLoadedSubsystem<XRImageTrackingSubsystem>();
 
-        public XRInputSubsystem inputSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRInputSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRInputSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRInputSubsystem inputSubsystem => GetLoadedSubsystem<XRInputSubsystem>();
 
-        public XRFaceSubsystem faceSubsystem
-        {
-            get
-            {
-                return GetLoadedSubsystem<XRFaceSubsystem>();
-            }
-        }
+        /// <summary>
+        /// The `XRFaceSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XRFaceSubsystem faceSubsystem => GetLoadedSubsystem<XRFaceSubsystem>();
 
+        /// <summary>
+        /// The `XREnvironmentProbeSubsystem` whose lifecycle is managed by this loader.
+        /// </summary>
+        public XREnvironmentProbeSubsystem environmentProbeSubsystem => GetLoadedSubsystem<XREnvironmentProbeSubsystem>();
+
+        /// <summary>
+        /// Initializes the loader.
+        /// </summary>
+        /// <returns>`True` if the session subsystem was successfully created, otherwise `false`.</returns>
         public override bool Initialize()
         {
 #if UNITY_ANDROID && !UNITY_EDITOR
@@ -100,6 +86,7 @@ namespace UnityEngine.XR.ARCore
             CreateSubsystem<XRImageTrackingSubsystemDescriptor, XRImageTrackingSubsystem>(s_ImageTrackingSubsystemDescriptors, "ARCore-ImageTracking");
             CreateSubsystem<XRInputSubsystemDescriptor, XRInputSubsystem>(s_InputSubsystemDescriptors, "ARCore-Input");
             CreateSubsystem<XRFaceSubsystemDescriptor, XRFaceSubsystem>(s_FaceSubsystemDescriptors, "ARCore-Face");
+            CreateSubsystem<XREnvironmentProbeSubsystemDescriptor, XREnvironmentProbeSubsystem>(s_EnvironmentProbeSubsystemDescriptors, "ARCore-EnvironmentProbe");
 
             if (sessionSubsystem == null)
             {
@@ -112,42 +99,28 @@ namespace UnityEngine.XR.ARCore
 #endif
         }
 
+        /// <summary>
+        /// Starts all subsystems.
+        /// </summary>
+        /// <returns>`True` if the subsystems were started, otherwise `false`.</returns>
         public override bool Start()
         {
-            var settings = GetSettings();
-            if (settings != null && settings.startAndStopSubsystems)
-            {
-                StartSubsystem<XRSessionSubsystem>();
-                StartSubsystem<XRCameraSubsystem>();
-                StartSubsystem<XRDepthSubsystem>();
-                StartSubsystem<XRPlaneSubsystem>();
-                StartSubsystem<XRAnchorSubsystem>();
-                StartSubsystem<XRRaycastSubsystem>();
-                StartSubsystem<XRImageTrackingSubsystem>();
-                StartSubsystem<XRInputSubsystem>();
-                StartSubsystem<XRFaceSubsystem>();
-            }
             return true;
         }
 
+        /// <summary>
+        /// Stops all subsystems.
+        /// </summary>
+        /// <returns>`True` if the subsystems were stopped, otherwise `false`.</returns>
         public override bool Stop()
         {
-            var settings = GetSettings();
-            if (settings != null && settings.startAndStopSubsystems)
-            {
-                StopSubsystem<XRSessionSubsystem>();
-                StopSubsystem<XRCameraSubsystem>();
-                StopSubsystem<XRDepthSubsystem>();
-                StopSubsystem<XRPlaneSubsystem>();
-                StopSubsystem<XRAnchorSubsystem>();
-                StopSubsystem<XRRaycastSubsystem>();
-                StopSubsystem<XRImageTrackingSubsystem>();
-                StopSubsystem<XRInputSubsystem>();
-                StopSubsystem<XRFaceSubsystem>();
-            }
             return true;
         }
 
+        /// <summary>
+        /// Destroys each subsystem.
+        /// </summary>
+        /// <returns>Always returns `true`.</returns>
         public override bool Deinitialize()
         {
 #if UNITY_ANDROID && !UNITY_EDITOR
@@ -160,21 +133,9 @@ namespace UnityEngine.XR.ARCore
             DestroySubsystem<XRImageTrackingSubsystem>();
             DestroySubsystem<XRInputSubsystem>();
             DestroySubsystem<XRFaceSubsystem>();
+            DestroySubsystem<XREnvironmentProbeSubsystem>();
 #endif
             return true;
-        }
-
-        ARCoreLoaderSettings GetSettings()
-        {
-            ARCoreLoaderSettings settings = null;
-            // When running in the Unity Editor, we have to load user's customization of configuration data directly from
-            // EditorBuildSettings. At runtime, we need to grab it from the static instance field instead.
-            #if UNITY_EDITOR
-            UnityEditor.EditorBuildSettings.TryGetConfigObject(ARCoreLoaderConstants.k_SettingsKey, out settings);
-            #else
-            settings = ARCoreLoaderSettings.s_RuntimeInstance;
-            #endif
-            return settings;
         }
     }
 }